refactor(alloy-eips): add more data validation to `SimpleCoder::decode_one()`

+++ crates/eips/src/eip4844/mod.rs
@@ -34,7 +34,7 @@ pub const BLS_MODULUS_BYTES: B256 =
 pub const BLS_MODULUS: U256 = U256::from_be_bytes(BLS_MODULUS_BYTES.0);
 
 /// Size a single field element in bytes.
-pub const FIELD_ELEMENT_BYTES: u64 = 32;
+pub const FIELD_ELEMENT_BYTES: usize = 32;
 
 /// How many field elements are stored in a single data blob.
 pub const FIELD_ELEMENTS_PER_BLOB: u64 = 4096;
